refactor: UserSession Dto로 이동
