+++ RTWWebServer/RTWWebServer.csproj
@@ -30,7 +30,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Authentication\" />
+        <Folder Include="Authentication\"/>
     </ItemGroup>
 
 </Project>
